+++ network/policy_updater/policy_updater.h
@@ -1,8 +1,8 @@
 /*
- * This file is part of the Frost distribution
- * (https://github.com/xainag/frost)
+ * This file is part of the IOTA Access Distribution
+ * (https://github.com/iotaledger/access)
  *
- * Copyright (c) 2019 XAIN AG.
+ * Copyright (c) 2020 IOTA Stiftung
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -21,7 +21,9 @@
 #define _POLICY_UPDATER_H_
 
 void PolicyUpdater_init();
-int PolicyUpdater_start();
-int PolicyUpdater_stop();
+
+void PolicyUpdater_get_policy(char *policy_id, char *policy_buff);
+
+unsigned int PolicyUpdater_get_policy_list(const char *policy_store_version, const char* device_id, char *policy_list, int *policy_list_len, int *new_policy_list_flag);
 
 #endif /* _POLICY_UPDATER_H_ */
